moved specific engine related files to their own directories/modules
