+++ simulariumio/physicell/physicell_converter.py
@@ -6,12 +6,13 @@ from typing import Any, Dict
 from pathlib import Path
 
 import numpy as np
-from .dep.pyMCDS import pyMCDS
+from ..dep.pyMCDS import pyMCDS
 
-from .converter import Converter
-from .data_objects import PhysicellData, AgentData
-from .exceptions import MissingDataError
-from .constants import VIZ_TYPE
+from ..converter import Converter
+from ..data_objects import AgentData
+from ..exceptions import MissingDataError
+from ..constants import VIZ_TYPE
+from .physicell_data import PhysicellData
 
 ###############################################################################
