+++ simulariumio/plot_readers/__init__.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
+from .plot_reader import PlotReader  # noqa: F401
 from .scatter_plot_reader import ScatterPlotReader  # noqa: F401
 from .histogram_plot_reader import HistogramPlotReader  # noqa: F401
