+++ simulariumio/plot_readers/plot_reader.py
