+++ simulariumio/plot_readers/scatter_plot_reader.py
