+++ simulariumio/readdy/__init__.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from .readdy_converter import ReaddyConverter  # noqa: F401
+from .readdy_data import ReaddyData  # noqa: F401
